+++ SystemServices/CHANGELOG.md
@@ -16,9 +16,13 @@ All notable changes to this RDK Service will be documented in this file.
 
 * For more details, refer to [versioning](https://github.com/rdkcentral/rdkservices#versioning) section under Main README.
 
-## [1.7.2] - 2023-10-18
-### Fixed
-- Fix for onFirmwareUpdateStateChange events with the same data.
+## [3.0.0] - 2023-11-09
+### Removed
+- Removed hasRebootBeenRequested method fixes
+
+## [2.0.0] - 2023-10-12
+### Removed
+- Removed the hardcoded implementation of hasRebootBeenRequested
 
 ## [1.7.1] - 2023-10-04
 ### Added